+++ lib/eskema.dart
@@ -1,4 +1,4 @@
-/// Eskema is a small, composable runtime validation library for Dart. It helps you validate 
+/// Eskema is a small, composable runtime validation library for Dart. It helps you validate
 /// dynamic values (JSON, Maps, Lists, primitives) with readable validators and clear error messages.
 ///
 /// **Use cases**:
@@ -15,7 +15,8 @@ export 'validators.dart';
 export 'validator.dart';
 export 'builder.dart';
 export 'result.dart';
-export 'extensions.dart';
 export 'transformers.dart';
 export 'expectation.dart';
 export 'error_format.dart';
+export 'extensions/base_extensions.dart';
+export 'extensions/operator_extensions.dart';